VMM: VMXAllTemplate.cpp.h: Assertion nit.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@100228 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMAll/VMXAllTemplate.cpp.h
@@ -10979,7 +10979,7 @@ HMVMX_EXIT_DECL vmxHCExitEptViolationNested(PVMCPUCC pVCpu, PVMXTRANSIENT pVmxTr
         }
 
         AssertMsg(Walk.fFailed & PGM_WALKFAIL_EPT_MISCONFIG,
-                  ("uErr=%#RX64 uExitQual=%#RX64 GCPhysNestedFault=%RGp GCPtrNestedFault=%RGv\n", uErr, uExitQual,
+                  ("uErr=%#RX64 uExitQual=%#RX64 GCPhysNestedFault=%RGp GCPtrNestedFault=%RGv\n", (uint64_t)uErr, uExitQual,
                    GCPhysNestedFault, GCPtrNestedFault));
         return IEMExecVmxVmexitEptMisconfig(pVCpu, pVmxTransient->uGuestPhysicalAddr, &ExitEventInfo);
     }
